requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0


